Move QuestionViewController to produciton
